update from WETH10

+++ contracts/IWETH10.sol
@@ -35,7 +35,7 @@ interface IWETH10 is IERC20, IERC2612 {
     /// The flash minted WETH10 is not backed by real Ether, but can be withdrawn as such up to the Ether balance of this contract.
     /// Arbitrary data can be passed as a bytes calldata parameter.
     /// Emits two {Transfer} events for minting and burning of the flash minted amount.
-    function flashMint(uint256 value, bytes calldata data) external;
+    function flashMint(uint112 value, bytes calldata data) external;
 
     /// @dev Burn `value` WETH10 token from caller account and withdraw matching ether to the same.
     /// Emits {Transfer} event to reflect WETH10 token burn of `value` WETH10 token to zero address from caller account.
